move registry
